+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 11.1
 - Choose Checkbox uncheck by default to prevent accidental creations in MassCreateAssignments.php
 - HTML display Courses list next to Assignment Types list in MassCreateAssignments.php
 - Set email for default admin user so he can reset his password in InstallDatabase.php
+- Prevent using App name, username, or email in the password in PasswordReset.php, FirstLogin.fnc.php, Inputs.php & Preferences.php
 
 Changes in 11.0.2
 -----------------